kafka/requirements.txt
kafka-python==2.0.2
requests==2.26.0
psycopg2-binary==2.9.1
python-dotenv==0.19.0